Rapport 0.2.0: fix what the pre-release audit found, and ship the voices
Release: make-release now packages Sound/Voice from the voice bank. It
fails if any line was never rendered, if a copied file differs, or if the
bank has uncommitted renders. Before this, a release would have played
every bark as a subtitle over silence. It also refuses to overwrite an
existing version's zip. The bank moves to its own private repository,
nested at voice/out, and this public repo stops tracking it. The version
is 0.2.0, plugin info carries the full number, and ApiVersion() returns
it to addons.

Records and narration wait for the scene to start: an affair noted for a
request is staged and written only then, and so is the Narrator's line.
An AAF refusal no longer leaves a permanent affair or a story about a
scene that never played.

Store: empty rows are pruned. The engine's seed is composed with earlier
movement instead of added to it, so order no longer matters. The partner
and incest flags follow the engine on every interaction. NaN is refused,
and a damaged save loads with sane values. A dead actor is not recorded
again when their scene ends. The seed and preview use the higher rank of
the two directions.

Safety: PersonaOf returns a copy, so an MCM reload can't free what a
reader holds. A malformed personas.json entry, or an MCM switch the json
lacks, no longer throws into F4SE at data ready. A borrowed voice from a
missing DLC falls back to the default instead of going silent. After a
bark the previous voice is restored, not cleared; this compiles against
F4SE's ObjectReference only. Listener ids above 0x7FFFFFFF survive the
trip to Papyrus. A new game clears in-flight state.

Scoring: only people count as onlookers (owner poll; children are
ignored entirely). The pair only discounts itself when it is inside the
onlooker radius. MCM settings apply even without scoring.json. A
candidate generation counter lets addons detect a torn read.

Narrator: repeated labels replace, marker labels carry the raw bond, the
couple flag and the decided-on score. It falls back to the addon's score
when the published offer moved. The two lines go in one order. "A first
time" only comes from a scene. Checked in game: a Sanctuary scene was
narrated at its start, and its parts sum to 1.75.

+++ src/Mailbox.cpp
@@ -328,7 +328,7 @@ namespace RP
 				return std::format("ERR {} <formid> <formid>{}", verb, verb == "bond" ? " [add <amount>]" : "");
 			}
 			if (verb == "rank") {
-				link.QueueOrder(Order{ Order::Kind::kRelation, a, std::to_string(b), "" });
+				link.QueueOrder(Order{ Order::Kind::kRelation, a, std::to_string(static_cast<std::int32_t>(b)), "" });
 				return std::format("OK queued - the engine's relationship between {:08X} and {:08X} lands in Rapport.log next poll", a, b);
 			}
 			auto& ledger = Ledger::GetSingleton();
